Fixes to mobile pop-out nav and other styles
CHANGES

()HTML

--> *Global
<div class="mobile-bar"> <img class="mobile-logo-top"
src="images/solliance-master.svg" 		 width="150"> </div>

—-Added this line of HTML which somehow was omitted from every HTML
page. This contains the mobile 	header and the logo that displays on
mobile devices.

--> index.html
Changed location in html of button under “we can help with your web
and mobile strategy” section. It 	was not centering because of where it
was placed in the hierarchy. Also added some crude formatting to
account for spacing. Will fix more appropriately later when CSS is
refactored.

()CLASSES

--> style.css / mobile.css / bootstrap.vertical-tabs.css
Edits made to multiple lines to fix issue with mobile pop-out menu
height
Edits made to tech.html page tabs to fix some alignment issues
Added a max width to image on the tech.html page that was causing
the 			 canvas to extend beyond 100%